// Servers/domain.layer/interfaces/i.control.ts
import { ISubcontrol } from "./i.subcontrol";

export type IControl = {
  id?: number; // auto generated by database
  title: string; // gets assigned from the structure
  description: string; // gets assigned from the structure
  order_no?: number; // gets assigned from the structure
  status?: "Waiting" | "In progress" | "Done"; // won't get any values, will be filled by user
  approver?: number; // won't get any values, will be filled by user
  risk_review?: "Acceptable risk" | "Residual risk" | "Unacceptable risk"; // won't get any values, will be filled by user
  owner?: number; // won't get any values, will be filled by user
  reviewer?: number; // won't get any values, will be filled by user
  due_date?: Date; // won't get any values, will be filled by user
  implementation_details?: string; // won't get any values, will be filled by user
  control_category_id: number; // when control category is created, its id will be stored and assign here as FK
  numberOfSubcontrols?: number;
  numberOfDoneSubcontrols?: number;
  subControls?: ISubcontrol[];
  created_at?: Date;
};